Brain damage keyboarding
The last 18 hours of Reddit Pluribus audience was people raging in out-group hate.

There is nothing wrong with the writing and language of Carol Sagan's last few books. Science as a Candle, Pale Blue Dot books but there is exactly what is mentioned in Science as a Candle going on with 10-second sound bite media. People are attracted to "least common denominator" programming. Vince Gillian sees it, and he can surf it like almost nobody else. But the problem is that one Bernie Sanders here or there doesn't cut it. In this GitHub there is a quote from Sagan too - that a few truths dumped upon here or there isn't dealing with the nation of suckers problem.