Hell is not a real place underground like a hotel for punishment.
Hell is the state of mind of Reddit audiences that wreck the nation with their attraction to "Operation Matthew 4:19" Trolling operation Russia and Trump originated.

Reddit cult of images can't face up to the November 7, 2013 private jet trip to Moscow that Trump got his final training on showcasing himself with girls and women to attract a Reddit fandom that never gets enough Trump bullshit.

Reddit society is the audience of Trump. People do not halt entertainment and seriously call for removal of Entertainment. The Reddit fanatics of Trump seek constant stimulation to worship the images and icons Putin and Trump provide. Leadership by images of stardom and lifestyle of the Putin buddies. Surkov won.