Luke Skywalker Ranch lessons for Star Wars light-saber phallic penis men
CAMPBELL: and our own Western subjugation of the female is a function of biblical thinking.

BILL MOYERS: Because when you substitute the male for the female, you get a different psychology, a different cultural bias. And it’s permissible in your culture to do what your gods do, so you just  —

CAMPBELL: That’s exactly it. I would see three situations here. First, the early one of the Goddess, when the male is hardly a significant divinity. Then the reverse, when the male takes over her role. And finally, then, the classical stage, where the two are in interaction—as they are also, for example, in India.