Expound Public Dreams of Dublin
Rock-a-billy BILL MOYERS: You talk about mythology existing here and now in dreamtime. What is dreamtime?

CAMPBELL: This is the time you get into when you go to sleep and have a dream that talks about permanent conditions within your own psyche as they relate to the temporal conditions of your life right now.

MOYERS: Explain that.

CAMPBELL: For example, you may be worried about whether you are going to pass an exam. Then you have a dream of some kind of failure, and you find that failure will be associated with many other failures in your life. They are all piled up together there. Freud says even the most fully expounded dream is not really fully expounded. The dream is an inexhaustible source of spiritual information about yourself.

Now the level of dream of “Will I pass the exam?” or “Should I marry this girl?”—that is purely personal. But, on another level, the problem of passing an exam is not simply a personal problem. Everyone has to pass a threshold of some kind. That is an archetypal thing. So there is a basic mythological theme there even though it is a personal dream. These two levels—the personal aspect and then the big general problem of which the person’s problem is a local example—are found in all cultures. For example, everyone has the problem of facing death. This is a standard mystery.


MOYERS: What do we learn from our dreams?

CAMPBELL: You learn about yourself.

MOYERS: How do we pay attention to our dreams?