What a cesspool of people who reply on Reddit and have no interest in serious discussion of USA POTUS  Seal symbol
Vince Gilligan: Climbing out of Social Media Cesspool

Writers Guild of America    
Vince Gilligan speech    
February 15, 2025     
https://youtu.be/TkteSxVIM1A?t=648   

10:48    
there's no simple answer for how *we* might change that. Though, if there were, I bet it would have something to do with climbing up out of the cesspool that is social media and [actually listening](https://youtu.be/QkF3oxziUI4?t=427) to each other

11:07        
But I don't know how to make that happen. So, I'll stick to my wheelhouse as
a writer speaking to a room full of writers. I have a proposal.