The caching in of gift and talent is for ideals and standing up to "cesspool" of social media
George Lucas filming audience interviews for the book "Power of Myth" in 1985, 1986, 1987 is more what I think Vince Gilligan is doing. This is cashing in bad-guy blood money storytelling.

the February 15, 2025 WGA speech says so much on the crisis of social media cesspools

____

Writers Guild of America
Vince Gilligan speech
February 15, 2025
https://youtu.be/TkteSxVIM1A?t=648

10:48
there's no simple answer for how we might change that. Though, if there were, I bet it would have something to do with climbing up out of the cesspool that is social media and actually listening to each other

11:07
But I don't know how to make that happen. So, I'll stick to my wheelhouse. As a writer speaking to a room full of writers, I have a proposal...